# swagger.yaml
openapi: 3.0.0
info:
  title: Zinda Learn REST API Backend
  version: 1.0.0
  description: API Documentation for MERN LMS Portal backend. Serving student profiles, teachers settings, courses registration, live classes stream endpoints, OTP logins, and payment integrations.
servers:
  - url: http://localhost:5005
    description: Development server
...
paths:
...
  /api/instructor/reviews:
    get:
      tags:
        - Teachers
      summary: Get reviews
      description: Retrieve reviews with user reference detailed info.
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Pagination page number
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
          description: Number of elements per page
        - name: search
          in: query
          schema:
            type: string
          description: Text string filter matching titles/descriptions
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Operation executed successfully
                  data:
                    type: object
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/instructor/students:
    get:
      tags:
        - Teachers
      summary: GET endpoint for students
      description: Performs operation on /api/instructor/students using GET method.
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Pagination page number
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
          description: Number of elements per page
        - name: search
          in: query
          schema:
            type: string
          description: Text string filter matching titles/descriptions
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Operation executed successfully
                  data:
                    type: object
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Error occurred
...